moved styles and components into a common directory
seems more sensible this way, particularly when considering the upcoming
addition of client-side JavaScript
