+++ lib/components/layout/styles/_footer.scss
