+++ lib/components/layout/styles/_header.scss
