+++ lib/images/faucet-logotype.gif
