+++ lib/images/faucet-logotype.png
