+++ lib/styles/settings/_colors.scss
