+++ package.json
@@ -1,7 +1,7 @@
 {
 	"scripts": {
 		"start": "npm-run-all compile --parallel server compile:watch build:watch",
-		"test": "eslint --cache --ext .js --ext .jsx *.js components views && echo ✓",
+		"test": "eslint --cache --ext .js --ext .jsx *.js lib/components views && echo ✓",
 		"server": "servedir ./dist/site",
 		"build": "node ./index.js",
 		"build:watch": "nodemon --delay 1 -w dist/views.js -w content -e md -x 'node ./index.js' < /dev/null",
